Changes : - Fixed torrent creation bug (this is a really moronic one ... , but the compiler should really give a warning when your passing an unsigned int variable to an unsigned char argument) - Moved chunk selection to separate class ChunkSelector - Added priority and exclude booleans to Chunk - Added functions to include, exclude or prioritise ranges of chunks in ChunkManager - ChunkSelector now takes into account priority and exclusion - Added a compare function in KTorrentViewItem
svn path=/trunk/extragear/network/ktorrent/; revision=439628

+++ libtorrent/bencoder.cpp
@@ -28,7 +28,7 @@ namespace bt
 	{
 	}
 
-	void BEncoderFileOutput::write(const char* str,Uint8 len)
+	void BEncoderFileOutput::write(const char* str,Uint32 len)
 	{
 		if (!fptr)
 			return;
